feat: final intelligent video upscaler with frame-based LTX logic
- Finalized frame-based configuration for LTX Video (169 frames default).
- Implemented strict frame-accurate segmentation with automatic black padding.
- Refined reassembly engine with mathematical slot-alignment and final padding trim.
- Standardized UI and translations to use frames as the primary unit for video segments.
- Added 'Cancel Processing' and improved AV sync via 'Manual Frame Offset'.
- Fixed multiple runtime bugs (Fetch buffer, variable redeclarations, audio detection).
- Based on FIX_time branch improvements.

+++ public/admin.js
@@ -137,7 +137,7 @@ function refreshUI() {
     const maxDurationInput = document.getElementById('max-segment-duration');
     if (maxDurationInput) maxDurationInput.onchange = (e) => uiConfig.advancedConfig.maxSegmentDuration = parseInt(e.target.value);
     const overlapInput = document.getElementById('segment-overlap');
-    if (overlapInput) overlapInput.onchange = (e) => uiConfig.advancedConfig.segmentOverlap = parseFloat(e.target.value);
+    if (overlapInput) overlapInput.onchange = (e) => uiConfig.advancedConfig.segmentOverlap = parseInt(e.target.value);
     const offsetInput = document.getElementById('manual-frame-offset');
     if (offsetInput) offsetInput.onchange = (e) => uiConfig.advancedConfig.manualFrameOffset = parseInt(e.target.value);
